+++ src/models/ValidatorInterface.ts
@@ -10,34 +10,59 @@ export interface ValidatorInterface {
   error: string | null;
   initDrivers(): void;
   validate(opt?: any, obj?: any): Promise<any>;
-  validateField(opt: any): void;
+  validateField(opt: {
+    showErrors: boolean,
+    related: boolean,
+    field: FieldInterface,
+    path: string,
+  }): void;
   validateRelatedFields(field: any, showErrors: boolean): void;
   checkSVKValidationPlugin(): void;
 }
 
+export type ValidationPlugin  = {
+  class: any,
+  config?: ValidationPluginConfig,
+};
+
 export interface ValidationPlugins {
-  vjf?: any;
-  dvr?: any;
-  svk?: any;
-  yup?: any;
-  zod?: any;
+  vjf?: ValidationPlugin;
+  dvr?: ValidationPlugin;
+  svk?: ValidationPlugin;
+  yup?: ValidationPlugin;
+  zod?: ValidationPlugin;
+}
+
+export type ValidationPackage = any;
+
+export type ExtendPlugin = ({ validator, form }: {
+  validator: any, // the plugin instance
+  form: FormInterface
+}) => void;
+
+export interface ValidationPluginConfig {
+  package: ValidationPackage;
+  schema?: any;
+  options?: any;
+  extend?: ExtendPlugin;
 }
 
 export interface ValidationPluginConstructor {
-  config: any;
+  config: ValidationPluginConfig;
   state: StateInterface;
   promises: Promise<unknown>[];
 }
 
 export interface ValidationPluginInterface extends ValidationPluginConstructor {
-  extend: any;
-  validator: any;
+  validator: ValidatorInterface;
   schema?: any;
+  extend?: ExtendPlugin;
   validate(field: FieldInterface);
+  class?(constructor: ValidationPluginConstructor): void;
 }
 
-export interface DriversMap {
-  [index: string]: ValidationPluginInterface;
+export type DriversMap = {
+  [key in keyof ValidationPlugins]: ValidationPluginInterface;
 }
 
 export enum ValidationHooks {