+++ src/validators/DVR.ts
@@ -1,5 +1,7 @@
 import _ from "lodash";
 import {
+  ValidationPlugin,
+  ValidationPluginConfig,
   ValidationPluginConstructor,
   ValidationPluginInterface,
 } from "../models/ValidatorInterface";
@@ -27,14 +29,14 @@ export class DVR implements ValidationPluginInterface {
   validator = null;
 
   constructor({
-    config = {},
+    config,
     state = null,
     promises = [],
   }: ValidationPluginConstructor) {
     this.state = state;
     this.promises = promises;
-    this.extend = config.extend;
-    this.validator = config.package || config;
+    this.extend = config?.extend;
+    this.validator = config.package;
     this.extendValidator();
   }
 
@@ -149,7 +151,7 @@ export class DVR implements ValidationPluginInterface {
   }
 }
 
-export default (config?: any) => ({
+export default (config?: ValidationPluginConfig): ValidationPlugin => ({
   class: DVR,
   config,
 });