+++ tests/data/forms/fixes/form.n.ts
@@ -38,7 +38,7 @@ const rules = {
 };
 
 const plugins: ValidationPlugins = {
-  dvr: dvr(validatorjs),
+  dvr: dvr({ package: validatorjs }),
 };
 
 class NewForm extends Form {}